Enforce C++ quality checks on pull requests
- Introduced a new GitHub Actions workflow (`cpp-quality-check.yml`) to run clang-format and clang-tidy checks on every pull request affecting C++ source files.
- Updated existing workflows to use recursive file discovery for clang-tidy and clang-format, ensuring all relevant .cpp and .h files are analyzed.
- Implemented error handling for cases where no source files are found, preventing silent successes.
- Modified several source files to resolve clang-tidy findings, including:
  - Changed parameter types to use const references in `MainComponent` and `MainComponentWorkspaceDrag`.
  - Added static_casts to ensure proper type conversions in `HarmonicAnalyzerComponent` and `FeedbackComponent`.
- Updated documentation to reflect new quality gate processes and recursive file discovery behavior.

+++ src/features/feedback/FeedbackComponent.cpp
@@ -308,8 +308,10 @@ juce::String FeedbackComponent::captureApplicationScreenshot() const
     const auto scale = juce::jmin(
         1.0f, maximumWidth / static_cast<float>(capturedBounds.getWidth()),
         maximumHeight / static_cast<float>(capturedBounds.getHeight()));
-    const auto imageWidth = juce::jmax(1, juce::roundToInt(capturedBounds.getWidth() * scale));
-    const auto imageHeight = juce::jmax(1, juce::roundToInt(capturedBounds.getHeight() * scale));
+    const auto imageWidth =
+        juce::jmax(1, juce::roundToInt(static_cast<float>(capturedBounds.getWidth()) * scale));
+    const auto imageHeight =
+        juce::jmax(1, juce::roundToInt(static_cast<float>(capturedBounds.getHeight()) * scale));
     juce::Image image(juce::Image::RGB, imageWidth, imageHeight, true);
     juce::Graphics graphics(image);
     graphics.fillAll(juce::Colours::black);
@@ -319,8 +321,10 @@ juce::String FeedbackComponent::captureApplicationScreenshot() const
         const auto snapshot =
             window->createComponentSnapshot(window->getLocalBounds(), true, scale);
         const auto screenBounds = window->getScreenBounds();
-        const auto x = juce::roundToInt((screenBounds.getX() - capturedBounds.getX()) * scale);
-        const auto y = juce::roundToInt((screenBounds.getY() - capturedBounds.getY()) * scale);
+        const auto x = juce::roundToInt(
+            static_cast<float>(screenBounds.getX() - capturedBounds.getX()) * scale);
+        const auto y = juce::roundToInt(
+            static_cast<float>(screenBounds.getY() - capturedBounds.getY()) * scale);
         graphics.drawImageAt(snapshot, x, y);
     }
